feat(settings): clean up functionality extraction
