+++ webview-ui/src/components/settings/SettingsView.tsx
@@ -18,7 +18,6 @@ import {
 	Database,
 	SquareTerminal,
 	FlaskConical,
-	Pencil,
 	AlertTriangle,
 	Globe,
 	Info,
@@ -66,7 +65,6 @@ import { LanguageSettings } from "./LanguageSettings"
 import { About } from "./About"
 import { Section } from "./Section"
 import PromptsSettings from "./PromptsSettings"
-import { FileEditingOptions } from "./FileEditingOptions"
 import { cn } from "@/lib/utils"
 
 export const settingsTabsContainer = "flex flex-1 overflow-hidden [&.narrow_.tab-label]:hidden"
@@ -82,7 +80,6 @@ export interface SettingsViewRef {
 
 const sectionNames = [
 	"providers",
-	"fileEditing",
 	"autoApprove",
 	"browser",
 	"checkpoints",
@@ -144,7 +141,6 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 		browserViewportSize,
 		enableCheckpoints,
 		diffEnabled,
-		fileBasedEditing,
 		experiments,
 		fuzzyMatchThreshold,
 		maxOpenTabsContext,
@@ -300,7 +296,6 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 			vscode.postMessage({ type: "ttsSpeed", value: ttsSpeed })
 			vscode.postMessage({ type: "soundVolume", value: soundVolume })
 			vscode.postMessage({ type: "diffEnabled", bool: diffEnabled })
-			vscode.postMessage({ type: "fileBasedEditing", bool: fileBasedEditing })
 			vscode.postMessage({ type: "enableCheckpoints", bool: enableCheckpoints })
 			vscode.postMessage({ type: "browserViewportSize", text: browserViewportSize })
 			vscode.postMessage({ type: "remoteBrowserHost", text: remoteBrowserHost })
@@ -412,7 +407,6 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 	const sections: { id: SectionName; icon: LucideIcon }[] = useMemo(
 		() => [
 			{ id: "providers", icon: Webhook },
-			{ id: "fileEditing", icon: Pencil },
 			{ id: "autoApprove", icon: CheckCheck },
 			{ id: "browser", icon: SquareMousePointer },
 			{ id: "checkpoints", icon: GitBranch },
@@ -605,25 +599,6 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 									setErrorMessage={setErrorMessage}
 								/>
 							</Section>
-						</div>
-					)}
-
-					{/* File Editing Section */}
-					{activeTab === "fileEditing" && (
-						<div>
-							<SectionHeader>
-								<div className="flex items-center gap-2">
-									<Pencil className="w-4" />
-									<div>{t("settings:sections.fileEditing")}</div>
-								</div>
-							</SectionHeader>
-
-							<FileEditingOptions
-								diffEnabled={diffEnabled}
-								fuzzyMatchThreshold={fuzzyMatchThreshold}
-								fileBasedEditing={fileBasedEditing}
-								onChange={setCachedStateField}
-							/>
 						</div>
 					)}
