+++ webview-ui/src/components/settings/__tests__/ApiOptions.spec.tsx
@@ -157,7 +157,7 @@ vi.mock("../RateLimitSecondsControl", () => ({
 }))
 
 // Mock DiffSettingsControl for tests
-vi.mock("../FileEditingOptions", () => ({
+vi.mock("../DiffSettingsControl", () => ({
 	DiffSettingsControl: ({ diffEnabled, fuzzyMatchThreshold, onChange }: any) => (
 		<div data-testid="diff-settings-control">
 			<label>
@@ -284,9 +284,15 @@ const renderApiOptions = (props: Partial<ApiOptionsProps> = {}) => {
 }
 
 describe("ApiOptions", () => {
-	it("shows temperature and rate limit controls by default", () => {
-		renderApiOptions()
+	it("shows diff settings, temperature and rate limit controls by default", () => {
+		renderApiOptions({
+			apiConfiguration: {
+				diffEnabled: true,
+				fuzzyMatchThreshold: 0.95,
+			},
+		})
 		// Check for DiffSettingsControl by looking for text content
+		expect(screen.getByText(/enable editing through diffs/i)).toBeInTheDocument()
 		expect(screen.getByTestId("temperature-control")).toBeInTheDocument()
 		expect(screen.getByTestId("rate-limit-seconds-control")).toBeInTheDocument()
 	})
